# src/main/resources/META-INF/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "ObraSync API",
    "version": "1.0.0",
    "description": "API de vistorias. ADMIN gerencia; ENGENHEIRO cria e edita as próprias; FISCAL consulta e altera status."
  },
  "servers": [
    {
      "url": "/obrasync/api"
    }
  ],
  "paths": {
...
    "/vistorias": {
      "get": {
        "summary": "Filtra e pagina no PostgreSQL",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Página de vistorias",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Pagina"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistência",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "description": "Nome do enum de status",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tipo",
            "in": "query",
            "required": false,
            "description": "Nome do enum de tipo",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "obra",
            "in": "query",
            "required": false,
            "description": "Trecho do nome da obra, sem distinção de caixa",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "responsavel",
            "in": "query",
            "required": false,
            "description": "Trecho do nome do responsável",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataInicial",
            "in": "query",
            "required": false,
            "description": "Data inclusiva yyyy-MM-dd",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "dataFinal",
            "in": "query",
            "required": false,
            "description": "Data inclusiva yyyy-MM-dd",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "description": "Página iniciando em zero",
            "schema": {
              "type": "integer",
              "default": 0,
              "minimum": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "description": "Itens por página",
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            }
          }
        ]
      },
      "post": {
        "summary": "Cria vistoria; engenheiro deve ser o responsável",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Criada; Location identifica o recurso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vistoria"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistência",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VistoriaRequest"
              }
            }
          }
        }
      }
    },
    "/vistorias/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Identificador da vistoria",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "get": {
        "summary": "Consulta uma vistoria",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Encontrada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vistoria"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistência",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Substitui os campos editáveis; mantém evidências",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Atualizada",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vistoria"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistência",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VistoriaRequest"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Exclui vistoria e evidências; somente ADMIN",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "204": {
            "description": "Excluída"
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistência",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
    "/vistorias/{id}/status": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Identificador da vistoria",
          "schema": {
            "type": "integer"
          }
        }
      ],
      "patch": {
        "summary": "Altera status; ADMIN ou FISCAL",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Status atualizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vistoria"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistência",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "APROVADA",
                      "PENDENTE",
                      "REPROVADA",
                      "EM_ANDAMENTO"
                    ]
                  }
                }
              }
            }
          }
        }
      }
    },
    "/vistorias/resumo": {
      "get": {
        "summary": "Métricas de todas as vistorias",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Resumo",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Resumo"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "401": {
            "description": "Token ausente ou inválido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "404": {
            "description": "Recurso não encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de persistência",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Erro"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Erro": {
        "type": "object",
        "properties": {
          "status": {
            "type": "integer"
          },
          "mensagem": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "VistoriaRequest": {
        "type": "object",
        "required": [
          "obraId",
          "responsavelId",
          "tipo"
        ],
        "properties": {
          "obraId": {
            "type": "integer",
            "format": "int64"
          },
          "responsavelId": {
            "type": "integer",
            "format": "int64"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "ESTRUTURAL",
              "ELETRICA",
              "HIDRAULICA",
              "ALVENARIA",
              "ACABAMENTO",
              "SEGURANCA_TRABALHO"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "APROVADA",
              "PENDENTE",
              "REPROVADA",
              "EM_ANDAMENTO"
            ]
          },
          "dataVistoria": {
            "type": "string",
            "format": "date"
          },
          "localizacao": {
            "type": "string",
            "maxLength": 150
          },
          "observacoes": {
            "type": "string"
          }
        },
        "example": {
          "obraId": 1,
          "responsavelId": 2,
          "tipo": "ELETRICA",
          "status": "PENDENTE",
          "dataVistoria": "2026-09-23",
          "localizacao": "Bloco A"
        }
      },
      "Vistoria": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "obraId": {
            "type": "integer",
            "format": "int64"
          },
          "responsavelId": {
            "type": "integer",
            "format": "int64"
          },
          "tipo": {
            "type": "string",
            "enum": [
              "ESTRUTURAL",
              "ELETRICA",
              "HIDRAULICA",
              "ALVENARIA",
              "ACABAMENTO",
              "SEGURANCA_TRABALHO"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "APROVADA",
              "PENDENTE",
              "REPROVADA",
              "EM_ANDAMENTO"
            ]
          },
          "dataVistoria": {
            "type": "string",
            "format": "date"
          },
          "localizacao": {
            "type": "string",
            "maxLength": 150
          },
          "observacoes": {
            "type": "string"
          },
          "obra": {
            "type": "string"
          },
          "responsavel": {
            "type": "string"
          }
        }
      },
      "Pagina": {
        "type": "object",
        "properties": {
          "itens": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Vistoria"
            }
          },
          "page": {
            "type": "integer"
          },
          "size": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "totalPages": {
            "type": "integer"
          }
        }
      },
      "Resumo": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer"
          },
          "aprovadas": {
            "type": "integer"
          },
          "pendentes": {
            "type": "integer"
          },
          "reprovadas": {
            "type": "integer"
          },
          "emAndamento": {
            "type": "integer"
          },
          "taxaAprovacao": {
            "type": "number"
          }
        }
      },
...
    }
  }
}